@@ -9,7 +9,7 @@
 	<parent>
 		<groupId>org.oxerr</groupId>
 		<artifactId>oxerr-parent</artifactId>
-		<version>1.6.2</version>
+		<version>1.6.4-SNAPSHOT</version>
 	</parent>
 	<groupId>org.oxerr.seatgeek.client</groupId>
 	<artifactId>seatgeek-client</artifactId>
@@ -22,8 +22,9 @@
 	<url>https://www.oxerr.org/seatgeek-client</url>
 	<inceptionYear>2023</inceptionYear>
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<log4j.version>2.20.0</log4j.version>
 		<ticket-inventory-support.version>0.0.2</ticket-inventory-support.version>
 	</properties>
 	<modules>
@@ -72,7 +73,7 @@
 			<dependency>
 				<groupId>jakarta.ws.rs</groupId>
 				<artifactId>jakarta.ws.rs-api</artifactId>
-				<version>2.1.6</version>
+				<version>3.1.0</version>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.commons</groupId>
@@ -82,17 +83,17 @@
 			<dependency>
 				<groupId>org.apache.logging.log4j</groupId>
 				<artifactId>log4j-api</artifactId>
-				<version>2.20.0</version>
+				<version>${log4j.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.logging.log4j</groupId>
 				<artifactId>log4j-core</artifactId>
-				<version>2.20.0</version>
+				<version>${log4j.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.logging.log4j</groupId>
 				<artifactId>log4j-slf4j2-impl</artifactId>
-				<version>2.20.0</version>
+				<version>${log4j.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>org.junit.jupiter</groupId>
@@ -102,12 +103,12 @@
 			<dependency>
 				<groupId>org.oxerr.rescu.ext</groupId>
 				<artifactId>rescu-ext-singleton</artifactId>
-				<version>0.0.10</version>
+				<version>0.0.11</version>
 			</dependency>
 			<dependency>
 				<groupId>org.oxerr.rescu.ext</groupId>
 				<artifactId>rescu-ext-rate-limiter-core</artifactId>
-				<version>0.0.10</version>
+				<version>0.0.11</version>
 			</dependency>
 			<dependency>
 				<groupId>org.oxerr.ticket.inventory.support</groupId>
@@ -122,25 +123,25 @@
 			<dependency>
 				<groupId>org.redisson</groupId>
 				<artifactId>redisson</artifactId>
-				<version>3.21.3</version>
+				<version>3.22.1</version>
 			</dependency>
 			<dependency>
 				<groupId>com.fasterxml.jackson.core</groupId>
 				<artifactId>jackson-databind</artifactId>
-				<version>2.15.1</version>
+				<version>2.15.2</version>
 			</dependency>
 
 			<!-- Java 8 Date/time -->
 			<dependency>
 				<groupId>com.fasterxml.jackson.datatype</groupId>
 				<artifactId>jackson-datatype-jsr310</artifactId>
-				<version>2.15.1</version>
+				<version>2.15.2</version>
 			</dependency>
 
 			<dependency>
 				<groupId>com.github.mmazi</groupId>
 				<artifactId>rescu</artifactId>
-				<version>2.1.0</version>
+				<version>3.0</version>
 			</dependency>
 			<dependency>
 				<groupId>com.google.code.findbugs</groupId>
